replace django4-cron library with management tasks
