+++ requirements.txt
@@ -30,13 +30,6 @@ charset-normalizer==3.2.0
 cryptography==41.0.3
     # via moto
 django==4.2.4
-    # via
-    #   -r requirements.in
-    #   django-common-helpers
-    #   django4-cron
-django-common-helpers==0.9.2
-    # via django4-cron
-django4-cron==0.5.1
     # via -r requirements.in
 idna==3.4
     # via requests
